Many things
I added Dotkid, Vkid, Telekid, and Lunarkid fairly cleanly. I also attempted to add Linekid, but it's a massive bruh OMEGALUL moment since the way the collision list is calculated needs to be changed.

Made the bow a drawn sprite in the player instead of a separate object

Gave blocks tangibility (Should be optimized / improved on later probably)

I added ladders quite cleanly (There is a bug with them, but I'll work it out later).

Added dynamic god mode.

Made the information overlay no longer a debug thing, and added much more information to it (Some of the information is still exclusive to debug mode, but most of it is freely accessible)

Made bullets smaller

Changed trigger fields to look more similar to how they are in renex engine.

Made all sounds Stereo