Change'd
I fixed jump refreshers not giving your jump back properly by changing the clamp to an approach.

I added the NANG star from RBN engine. They are under the Refresher category in Misc.

The only thing remaining to be done from Jumple Refrenture is to fix the Slowmo refresher's jump effect. (It's supposed to zoom in on the player, but I can't figure out how to do it in GMS2 lol)

I think outside of like 3 small things and a few optimizations that should be made, the Ocean Wetventure gimmicks should be nearly perfect.

Added conveyors and elevators (They're a little broken because I need to fix how the on_slide variable is set. You might take a look at that to see a better way to do it that I can't).

Added slip blocks. They have the same issue that on_slide has, so also probably look at on_ice.